% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cover.R
\name{calculate_cover}
\alias{calculate_cover}
\title{Calculate total cover per species in vegetation transects}
\usage{
calculate_cover(df, split.cols, first.spcol, last.spcol = ncol(df), tr.length,
  bare = FALSE, precision, check.incremental = TRUE, prop = FALSE,
  long.format = FALSE, show.progress = TRUE)
}
\arguments{
\item{df}{A dataframe with one column per recorded species, plus additional columns
to identify transects (and possibly sections within transects, shrub or tree layers, etc).
For each species, introduce initial and ending distances every time the species appears along
the transect (see example).}

\item{split.cols}{A character vector with the names of the variables in \code{df}
that identify transects or any other structure of the data (e.g. sections within transects,
shrub or tree layers, etc.)}

\item{first.spcol}{Integer. Number of the first column containing species cover data.}

\item{last.spcol}{Integer. Number of the last column containing species cover data
(last column in \code{df} by default).}

\item{tr.length}{Numeric. Total transect length (e.g. in meters).}

\item{bare}{Logical. Calculate bare cover? Default is \code{TRUE}.}

\item{precision}{Numeric. Precision of species cover measurements. Only used to calculate
bare cover by exclusion.}

\item{check.incremental}{Logical (default is \code{TRUE}). Check that cover data are
sorted incrementally? Must be \code{TRUE} if calculating bare cover with this function.}

\item{prop}{Logical. If \code{TRUE}, results are expressed as proportion
of \code{tr.length} covered by each species. If \code{FALSE} (the default), results represent
linear absolute length covered by each species.}

\item{long.format}{Logical. Produce the output dataframe in long format? Default is wide format.
See \code{\link[tidyr]{gather}}.}

\item{show.progress}{Logical. Show progress bar?}
}
\value{
A dataframe with total cover values per species and transect (plus optionally bare ground).
}
\description{
Calculate total cover per species in each transect, layer within transect, etc.
Optionally, the function can also calculate length of bare ground (as ground not covered
by any plant species).
}
\examples{
df <- data.frame(transect = sort(rep(c(1,2,3), 20)),
                 section = rep(1, 60),
                 species1 = c(replicate(3, sort(round(runif(20, 0, 100), digits = 1)))),
                 species2 = c(sort(round(runif(8, 0, 100), digits = 1)), rep(NA, times = 60 - 8)))
head(df)
calculate_cover(df, split.cols = "transect", first.spcol = 3,
tr.length = 100, check.incremental = FALSE)
calculate_cover(df, split.cols = "transect", first.spcol = 3, tr.length = 100,
check.incremental = FALSE, long.format = TRUE)
}
\author{
F Rodriguez-Sanchez
}

